deal with various compiler warnings
Note that this includes various safety related changes for the multi ADD/GET macros.

+++ include/multi.h
@@ -401,7 +401,7 @@ struct CFILE;
 //
 
 // netplayer management
-#define NET_PLAYER_INDEX(np)	(np-Net_players)
+#define NET_PLAYER_INDEX(np)	(static_cast<int>(np-Net_players))
 #define NET_PLAYER_NUM(np)		(NET_PLAYER_INDEX(np))
 #define MY_NET_PLAYER_NUM		(NET_PLAYER_INDEX(Net_player))
